# PERF: quickly resolve inequalities for Quantities with different physical types