--- a/astropy/units/quantity.py
+++ b/astropy/units/quantity.py
@@ -1138,10 +1138,9 @@ class Quantity(np.ndarray):
     # DeprecationWarnings on any error, which is distracting, and does not
     # deal well with structured arrays (nor does the ufunc).
     def __eq__(self, other):
-        if (
-            hasattr(other, "unit")
-            and getattr(other.unit, "physical_type", True) != getattr(self.unit, "physical_type", False)
-        ):
+        if hasattr(other, "unit") and getattr(
+            other.unit, "physical_type", True
+        ) != getattr(self.unit, "physical_type", False):
             # hot path for flagrant inequality cases
             return False
         try:
@@ -1155,8 +1154,7 @@ class Quantity(np.ndarray):
     def __ne__(self, other):
         if (
             hasattr(other, "unit")
-            and hasattr(other.unit, "physical_type")
-            and (other.unit.physical_type) != self.unit.physical_type
+            and getattr(other.unit, "physical_type", True) != self.unit.physical_type
         ):
             # hot path for flagrant inequality cases
             return True