@mhvk 2024-03-28
@neutrinoceros - note that in your example (at least as written) the unit of `data2` is "g kg", so the unit is not dimensionally compatible (I expect this is an error with adding the unit to `data` later?). I think we should definitely prioritize comparisons with compatible units...