@mhvk 2024-03-28
p.s. Bit more generally, I think it would be best to focus on speeding up (if possible), `_to_own_unit()` (and hence `to_value()`) since those are used very commonly elsewhere as well.