@nstarman 2024-03-28
> running the test suite locally, I see that it breaks 3 tests, all for the same reason: redshifts' physical type doesn't compare equal to the dimensionless' physical type

Yes, this is a deliberate choice. Redshift has a physical meaning related to the distance. The primary reason for having a redshift unit is to track factors of redshift across a calculation and distinguish it from normal dimensionless numbers.