@neutrinoceros 2024-03-29
> note that in your example (at least as written) the unit of data2 is "g kg", so the unit is not dimensionally compatible (I expect this is an error with adding the unit to data later?).
> I think we should definitely prioritize comparisons with compatible units...

This may be a mistake, but I took it as is from the benchmark code:

https://github.com/astropy/astropy-benchmarks/blob/c54d814aef79930d5cf6036d89c7c16c16d7740e/benchmarks/units.py#L159-L164

I think this case is actually worth a benchmark, but we clearly need to have another one in the mix too. I'll open a PR today.

> I think it would be best to focus on speeding up (if possible), _to_own_unit() (and hence to_value()) since those are used very commonly elsewhere as well.

Perhaps I should have told the broader story in the PR description: this actually was my original goal, until I discovered that a lot of the logic to distinguish between compatible and incompatible units was built using `try/except` blocks, which are the root cause of the performance imbalance I'm revealing. I expect that refactoring this logic to not rely on exception catching would be very difficult, so I tried to work on a different level instead, hence this patch.

>  The primary reason for having a redshift unit is to track factors of redshift across a calculation and distinguish it from normal dimensionless numbers.

I'm slightly confused. Then why was it decided to allow redshift quantites to compare equal to dimensionless ones ? Isn't there a conflict between these two choices ?