@neutrinoceros 2024-03-29
Also, while away from keyboard, I wanted to know the tipping point between the reference implementation and my branch, in terms of the relative abundance of hits VS misses (where I define a hit as a comparison between equal objects and a miss as a comparison between unequal objects). The answer I found is about 4.5. In other words, when 18% (or more) of all comparisons being made are misses, my branch is overall faster. This seems low enough that I can phantom most applications would benefit from it.