@neutrinoceros 2024-03-29
And refactoring futher to skip `PhysicalType.__eq__` we get:
```
'data==data2' took 3.1 s (or 6.2 us/iteration)
'data!=data2' took 3.1 s (or 6.2 us/iteration)
'data==data3' took 3.8 s (or 7.6 us/iteration)
'data!=data3' took 3.8 s (or 7.6 us/iteration)
```

further moving the tipping point to ~32, meaning just 3% misses is sufficient to make the patch an overall gain.
With this version, the happy path is only degraded by about 9%, while the "unhappy" path is improved by more than 4x
Looks to me this is getting less controversial by the minute 😸