@mhvk 2024-03-29
@neutrinoceros - I think you overestimate the importance of unequal units in `Quantity.__eq__` - I would think that in almost all cases, that is a problem with user input and hence soon enough an exception will be raised.  I suspect a large fraction is comparisons with equivalent units where the resulting array of bool is used to index or so (though, given floats, that is perhaps more likely for greater than, etc., except for `== 0`. (Note that those code paths will do badly regardless with `__eq__` returning a plain `False` - we never properly followed numpy's lead in ensuring that `__eq__` always returns an array with broadcasted shape...).

One possible way to test this is to count results in the astropy test suite (though there again I'd be wary of tests that explicitly check wrong user input...)

Overall, I'm very wary of performance improvements that work by adding code. For instance, in this case, if python manages to make the except path faster, any `if`s beforehand will just make the code slower (and this is not so unrealistic: it is only a few versions ago that python ensured that the success path of a `try/except` has no cost).

More specific to this PR - it seems the main gain is from making unit comparisons faster - that seems very worthwhile on its own, but might have wider benefit if done in the units code itself. Though it is hard to be sure without doing a bit more detailed profiling.