@neutrinoceros 2024-03-29
I just realized that my approach was intrinsically biased towards unequal cases because of a mistake I made while reading/adapting the benchmark code 🤦🏻‍♂️ 
In actuality, the benchmark is really comparing quantities that are dimensionally compatible, I just broke this property as I extracted the code from a benchmark class and dropped the distinction between two variables (`self.data` and `data`).

So, thanks a lot for being so wary, it's clearly pushed me in the right direction of critical thinking 😅
I also agree that it'd be best to optimize `Unit.__eq__` instead of `Quantity.__eq__` if possible, since the former implies the latter. Anyway, now I see that this specific patch is probably not going anywhere, so I'll close it and start over. Thank the both of you for replying so promptly !